+++ examples/compression.rs
@@ -1,8 +1,7 @@
 //! Numeric flags similar to zip compression levels, accepts -1 to -9, produces usize
 
 use bpaf::{doc::Style, *};
-
-fn compression() -> impl Parser<usize> {
+fn compression() -> impl Parser<Output = usize> {
     any::<isize, _, _>("COMP", |x: isize| {
         if (-9..=-1).contains(&x) {
             Some(x.abs().try_into().unwrap())