+++ examples/many_comma_separated_args.rs
@@ -3,7 +3,7 @@ use bpaf::*;
 use std::str::FromStr;
 
 // --ports 1,2,3 --ports 4,5   => [1,2,3,4,5]
-fn args() -> impl Parser<Vec<u16>> {
+fn args() -> impl Parser<Output = Vec<u16>> {
     long("ports")
         .help("Comma separated list of ports")
         .argument::<String>("PORTS")