+++ examples/top_to_bottom.rs
@@ -30,14 +30,14 @@ fn main() {
     println!("{:#?}", opt);
 }
 // A flag, true if used in the command line. Can be required, this one is optional
-fn debug() -> impl Parser<bool> {
+fn debug() -> impl Parser<Output = bool> {
     short('d')
         .long("debug")
         .help("Activate debug mode")
         .switch()
 }
 // number of occurrences of the v/verbose flag capped at 3
-fn verbose() -> impl Parser<usize> {
+fn verbose() -> impl Parser<Output = usize> {
     short('v')
         .long("verbose")
         .help("Increase the verbosity\nYou can specify it up to 3 times\neither as -v -v -v or as -vvv")
@@ -48,7 +48,7 @@ fn verbose() -> impl Parser<usize> {
 }
 
 // an argument, parsed and with default value
-fn speed() -> impl Parser<f64> {
+fn speed() -> impl Parser<Output = f64> {
     short('s')
         .long("speed")
         .help("Set speed")
@@ -56,19 +56,19 @@ fn speed() -> impl Parser<f64> {
         .fallback(42.0)
 }
 
-fn output() -> impl Parser<PathBuf> {
+fn output() -> impl Parser<Output = PathBuf> {
     short('o')
         .long("output")
         .help("output file")
         .argument::<PathBuf>("OUTPUT")
 }
 
-// no magical name transmogrifications.
-fn nb_cars() -> impl Parser<u32> {
+// no magical name transmogrification.
+fn nb_cars() -> impl Parser<Output = u32> {
     short('n').long("nb-cars").argument::<u32>("N")
 }
 
-fn files_to_process() -> impl Parser<Vec<PathBuf>> {
+fn files_to_process() -> impl Parser<Output = Vec<PathBuf>> {
     short('f')
         .long("file")
         .help("File to process")